refactor: extract task run token verification helper
